Moves common macro functionality in MacroTrait. Also unifies macro tests and space correction
