+++ src/Illuminate/Support/Facades/Response.php
@@ -5,15 +5,11 @@ use Illuminate\Http\Response as IlluminateResponse;
 use Illuminate\Support\Contracts\ArrayableInterface;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Illuminate\Support\Traits\MacroTrait;
 
 class Response {
 
-	/**
-	 * An array of registered Response macros.
-	 *
-	 * @var array
-	 */
-	protected static $macros = array();
+	use MacroTrait;
 
 	/**
 	 * Return a new response from the application.
@@ -93,37 +89,6 @@ class Response {
 		}
 
 		return $response;
-	}
-
-	/**
-	 * Register a macro with the Response class.
-	 *
-	 * @param  string  $name
-	 * @param  callable  $callback
-	 * @return void
-	 */
-	public static function macro($name, $callback)
-	{
-		static::$macros[$name] = $callback;
-	}
-
-	/**
-	 * Handle dynamic calls into Response macros.
-	 *
-	 * @param  string  $method
-	 * @param  array  $parameters
-	 * @return mixed
-	 *
-	 * @throws \BadMethodCallException
-	 */
-	public static function __callStatic($method, $parameters)
-	{
-		if (isset(static::$macros[$method]))
-		{
-			return call_user_func_array(static::$macros[$method], $parameters);
-		}
-
-		throw new \BadMethodCallException("Call to undefined method $method");
 	}
 
 }
