+++ tests/Support/SupportStrTest.php
@@ -23,11 +23,13 @@ class SupportStrTest extends PHPUnit_Framework_TestCase {
 		$this->assertEquals(' Taylor...', Str::words(' Taylor Otwell ', 1));
 	}
 
+
 	public function testStringTitle()
 	{
 		$this->assertEquals('Jefferson Costella', Str::title('jefferson costella'));
 		$this->assertEquals('Jefferson Costella', Str::title('jefFErson coSTella'));
 	}
+
 
 	public function testStringWithoutWordsDoesntProduceError()
 	{
@@ -39,8 +41,8 @@ class SupportStrTest extends PHPUnit_Framework_TestCase {
 
 	public function testStringMacros()
 	{
-		Illuminate\Support\Str::macro(__CLASS__, function() { return 'foo'; });
-		$this->assertEquals('foo', Str::SupportStrTest());
+		Str::macro(__CLASS__, function() { return 'foo'; });
+		$this->assertEquals('foo', Str::{__CLASS__}());
 	}
